+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-06-19 10:42:48
-// Commit: feat(forms): add bi-directional event bus handler for forms
+// Updated at: 2026-06-19 22:07:18
+// Commit: docs(forms): update readme and API contracts for forms
 
 class formsComponentState {
-  final String commitRef = '2122';
-  final String lastMessage = 'feat(forms): add bi-directional event bus handler for forms';
-  final DateTime timestamp = DateTime.parse('2026-06-19T10:42:48+05:00');
+  final String commitRef = '2159';
+  final String lastMessage = 'docs(forms): update readme and API contracts for forms';
+  final DateTime timestamp = DateTime.parse('2026-06-19T22:07:18+05:00');
 }